vacancyViewHolder was extracted from inner class to separate class
